chore(space-wizard): rename and move files
